update: 1.data_process_gen.py 10.rkFlameGraph.sh

+++ 2.dataProc/1.data_process_gen.py
@@ -13,7 +13,7 @@ def loadData(fname):
 
     nums = []
     for line in lines:
-        cur = int(line.strip('\n'))
+        cur = float(line.strip('\n'))
         nums.append(cur)
 
     return nums
